+++ service/collector/cluster/nodepools.go
@@ -14,7 +14,7 @@ import (
 	expcapiv1beta1 "sigs.k8s.io/cluster-api/exp/api/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/giantswarm/azure-collector/v2/internal/capzcredentials"
+	"github.com/giantswarm/azure-collector/v3/internal/capzcredentials"
 )
 
 type NodePools struct {