+++ service/collector/set.go
@@ -6,7 +6,7 @@ import (
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 
-	"github.com/giantswarm/azure-collector/v2/service/collector/cluster"
+	"github.com/giantswarm/azure-collector/v3/service/collector/cluster"
 )
 
 const (