+++ service/collector/sp_expiration.go
@@ -9,8 +9,8 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/giantswarm/azure-collector/v2/client"
-	"github.com/giantswarm/azure-collector/v2/service/credential"
+	"github.com/giantswarm/azure-collector/v3/client"
+	"github.com/giantswarm/azure-collector/v3/service/credential"
 )
 
 const (