+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0] - 2023-04-18
+
 ### Changed
 
 - Use CtrlClient instead of typed clients.
@@ -178,7 +180,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/giantswarm/azure-collector/compare/v2.13.0...HEAD
+[Unreleased]: https://github.com/giantswarm/azure-collector/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/giantswarm/azure-collector/compare/v2.13.0...v3.0.0
 [2.13.0]: https://github.com/giantswarm/azure-collector/compare/v2.12.0...v2.13.0
 [2.12.0]: https://github.com/giantswarm/azure-collector/compare/v2.11.0...v2.12.0
 [2.11.0]: https://github.com/giantswarm/azure-collector/compare/v2.10.0...v2.11.0